requirements.txt
flask
pymongo
dnspython
flask-cors
python-dotenv
